minor tweaks

+++ install.sh
@@ -253,9 +253,9 @@ apt-get autoremove -y
 
 # Let's remove NGINX default website
 fuECHO "### Removing NGINX default website."
-rm /etc/nginx/sites-enabled/default  2>&1
-rm /etc/nginx/sites-available/default  2>&1
-rm /usr/share/nginx/html/index.html  2>&1
+[ -f /etc/nginx/sites-enabled ] && rm /etc/nginx/sites-enabled/default  
+[ -f /etc/nginx/sites-avaliable ] && rm /etc/nginx/sites-available/default  
+[ -f /usr/share/nginx/html/index.html ] && rm /usr/share/nginx/html/index.html  
 
 if [ -z ${noninteractive+x} ]; then
 	# Let's ask user for a password for the web user
@@ -309,7 +309,7 @@ pip install docker-compose==1.16.1
 fuECHO "### Installing elasticsearch curator."
 pip install elasticsearch-curator==5.2.0
 fuECHO "### Installing wetty."
-ln -s /usr/bin/nodejs /usr/bin/node 
+[ ! -f /usr/bin/node ] && ln -s /usr/bin/nodejs /usr/bin/node 
 npm install https://github.com/t3chn0m4g3/wetty -g 
 fuECHO "### Installing elasticsearch-dump."
 npm install https://github.com/t3chn0m4g3/elasticsearch-dump -g
